requirements.txt
# === requirements.txt ===
Flask
requests
Pillow
numpy
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
Werkzeug
google-cloud-storage
